Adds authorization to user actions
Prior to this change, is was not possible to retrieve the permission from a bookmark for AuthorizedUserTasks.

This change updates the AuthorizedUserTask activity to persist the permission in order to retrieve it when setting bookmarks.  This allows filtering all user actions to only those to which the user has the required permission.