Adds ability to reset and detach workflow instances
This change adds the ability to reset a workflow to the beginning, by running an alteration plan to remove bookmarks and schedule an activity.

This change also adds a new state, Detached, that represents what would happen if a workflow definition needed to remove a state.  The detached state does not have the ability to transition to a new state.
Instead, the reset workflow can be issued to reset the workflow.